Move Result.Unmarshal method to the json module, fix review comments
