+++ result/result.go
@@ -1,7 +1,6 @@
 package result
 
 import (
-	"encoding/json"
 	"fmt"
 )
 
@@ -99,17 +98,4 @@ func (r Result[T]) UnwrapErr() error {
 	}
 
 	return r.err
-}
-
-// UnmarshalJSON implements the [json.Unmarshaler] interface.
-// Values will be unmarshaled as [Ok] variants.
-func (r *Result[T]) UnmarshalJSON(data []byte) error {
-	var value T
-
-	if err := json.Unmarshal(data, &value); err != nil {
-		return err
-	}
-
-	*r = Ok(value)
-	return nil
 }
